Fix hang when connection drops. (#7)
* Fix hang when connection drops.

Added additional checks for connection after GetFTPAnswer calls throughout.  This results in faster identification of connection drop and elminates unnecessary timeouts that delay return back to main code.  InitFile closes the connection and updates the connection status if a BadPASV Answer.  This corrects the hang when the connection drops and tries to establish the connection to the Data Port.

* Added timeout to client connections.

https://github.com/ldab/ESP32_FTPClient/issues/6